Refactor subquery integrated tests (#4721)
