+++ shardingsphere-sql-parser/shardingsphere-sql-parser-dialect/shardingsphere-sql-parser-mysql/src/main/java/org/apache/shardingsphere/sql/parser/mysql/visitor/MySQLVisitor.java
@@ -287,9 +287,16 @@ public abstract class MySQLVisitor extends MySQLStatementBaseVisitor<ASTNode> {
         if (null != ctx.subquery()) {
             new SubquerySegment(ctx.getStart().getStartIndex(), ctx.getStop().getStopIndex(), (SelectStatement) visit(ctx.subquery()));
         }
-        ASTNode result = visit(ctx.predicate());
-        return result instanceof ColumnSegment
-                ? (ColumnSegment) result : new PredicateCompareRightValue(ctx.comparisonOperator().getText(), (ExpressionSegment) result);
+        ASTNode rightValue = visit(ctx.predicate());
+        return createPredicateRightValue(ctx, rightValue);
+    }
+    
+    private ASTNode createPredicateRightValue(final BooleanPrimaryContext ctx, final ASTNode rightValue) {
+        if (rightValue instanceof ColumnSegment) {
+            return rightValue;
+        }
+        return rightValue instanceof SubquerySegment ? new PredicateCompareRightValue(ctx.comparisonOperator().getText(), new SubqueryExpressionSegment((SubquerySegment) rightValue)) 
+                : new PredicateCompareRightValue(ctx.comparisonOperator().getText(), (ExpressionSegment) rightValue);
     }
     
     @Override
