+++ shardingsphere-sql-parser/shardingsphere-sql-parser-dialect/shardingsphere-sql-parser-postgresql/src/main/java/org/apache/shardingsphere/sql/parser/postgresql/visitor/PostgreSQLVisitor.java
@@ -271,9 +271,16 @@ public abstract class PostgreSQLVisitor extends PostgreSQLStatementBaseVisitor<A
         if (null != ctx.subquery()) {
             new SubquerySegment(ctx.getStart().getStartIndex(), ctx.getStop().getStopIndex(), (SelectStatement) visit(ctx.subquery()));
         }
-        ASTNode result = visit(ctx.predicate());
-        return result instanceof ColumnSegment
-                ? (ColumnSegment) result : new PredicateCompareRightValue(ctx.comparisonOperator().getText(), (ExpressionSegment) result);
+        ASTNode rightValue = visit(ctx.predicate());
+        return createPredicateRightValue(ctx, rightValue);
+    }
+    
+    private ASTNode createPredicateRightValue(final BooleanPrimaryContext ctx, final ASTNode rightValue) {
+        if (rightValue instanceof ColumnSegment) {
+            return rightValue;
+        }
+        return rightValue instanceof SubquerySegment ? new PredicateCompareRightValue(ctx.comparisonOperator().getText(), new SubqueryExpressionSegment((SubquerySegment) rightValue))
+                : new PredicateCompareRightValue(ctx.comparisonOperator().getText(), (ExpressionSegment) rightValue);
     }
     
     @Override
